+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+# 1.1.1
+- Warnings added for problematic variables such as "LET PRINT=1".
+- Focus to PROBLEM pane fixed.
+
+# 1.1.0
+- Execution of ZX81 P-files.
+
 # 1.0.0
-- First version
+- First version.